+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 
 -   Improved logging system to simplify the identification of issues in a production build
+-   Better support of workspace generation based on Package.swift as tuist can not generate workspace without .git folder set
+-   Better macro error parser
 
 ## 0.6.2 - 2025-12-25